Extracted scopes and query building into traits
